add star and comets
